xsrf

+++ www/app/controllers/page/test.c.php
@@ -44,49 +44,54 @@
 				
 				$answers = AnswerModel::loadAllForQuestion($this->db, $question->ival('belbin_question_id'));
 						
-				if (z::isPost()) {		
+				if (z::isPost()) {				
 				
-					$answered_score = 0;
-					$results = [];
-							
-					foreach ($answers as $answer) {
-						$question_score = z::getInt('answer_' . $answer->val('belbin_answer_id'), 0);
-						if ($question_score > 0) {
-							$answered_score += $question_score;
-							$result = new ResultModel($this->db);
-							$result->set('belbin_result_test_id', $test->ival('belbin_test_id'));
-							$result->set('belbin_result_question_id', $question->ival('belbin_question_id'));
-							$result->set('belbin_result_answer_id', $answer->ival('belbin_answer_id'));
-							$result->set('belbin_result_score', $question_score);
-							$results[] = $result;
-						}
-					}		
-
-					//z::debug($answered_score);		
+					if ($this->z->forms->verifyXSRFTokenHash('belbin_test_form', z::get('form_token'))) {
 					
-					if ($answered_score == TestModel::$score_per_question) {
-						ResultModel::deleteAllQuestionResults($this->db, $question->ival('belbin_question_id'), $test->ival('belbin_test_id'));
-						foreach ($results as $result) {				
-							$result->save();
-						}
-						$question = QuestionModel::loadNextQuestion($this->db, $question->ival('belbin_question_index'));			
-				
-						if (isset($question) && $question->is_loaded) {							
-							$answers = AnswerModel::loadAllForQuestion($this->db, $question->ival('belbin_question_id'));
-						} else {
-							if ($test->validateTestResults()) {
-								//test is ok - finish the test and redirect to result
-								$test->set('belbin_test_end_date', zSqlQuery::mysqlDatetime(time()));
-								$test->save();
-								$this->redirect(sprintf('default/default/result/%d', $test->ival('belbin_test_id')));
-							} else {
-								//test is not valid - redirect to last question and ask user to answer all questions
-								$question = $original_question;
-								$this->message('Není přidělen dostatečný počet bodů a test nelze uzavřít! Projděte si prosím všechny otázky znovu a ujistěte se, že jste u každé rozdělili požadovaný počet bodů.', 'error');
+						$answered_score = 0;
+						$results = [];
+								
+						foreach ($answers as $answer) {
+							$question_score = z::getInt('answer_' . $answer->val('belbin_answer_id'), 0);
+							if ($question_score > 0) {
+								$answered_score += $question_score;
+								$result = new ResultModel($this->db);
+								$result->set('belbin_result_test_id', $test->ival('belbin_test_id'));
+								$result->set('belbin_result_question_id', $question->ival('belbin_question_id'));
+								$result->set('belbin_result_answer_id', $answer->ival('belbin_answer_id'));
+								$result->set('belbin_result_score', $question_score);
+								$results[] = $result;
 							}
+						}		
+
+						//z::debug($answered_score);		
+						
+						if ($answered_score == TestModel::$score_per_question) {
+							ResultModel::deleteAllQuestionResults($this->db, $question->ival('belbin_question_id'), $test->ival('belbin_test_id'));
+							foreach ($results as $result) {				
+								$result->save();
+							}
+							$question = QuestionModel::loadNextQuestion($this->db, $question->ival('belbin_question_index'));			
+					
+							if (isset($question) && $question->is_loaded) {							
+								$answers = AnswerModel::loadAllForQuestion($this->db, $question->ival('belbin_question_id'));
+							} else {
+								if ($test->validateTestResults()) {
+									//test is ok - finish the test and redirect to result
+									$test->set('belbin_test_end_date', zSqlQuery::mysqlDatetime(time()));
+									$test->save();
+									$this->redirect(sprintf('default/default/result/%d', $test->ival('belbin_test_id')));
+								} else {
+									//test is not valid - redirect to last question and ask user to answer all questions
+									$question = $original_question;
+									$this->message('Není přidělen dostatečný počet bodů a test nelze uzavřít! Projděte si prosím všechny otázky znovu a ujistěte se, že jste u každé rozdělili požadovaný počet bodů.', 'error');
+								}
+							}
+						} else {
+							$this->z->messages->add($this->t('Distribute exactly %d points!', TestModel::$score_per_question), 'error');
 						}
 					} else {
-						$this->z->messages->add($this->t('Distribute exactly %d points!', TestModel::$score_per_question), 'error');
+						$this->z->messages->add('Byl detekován pokus o opakované odeslání formuláře! Data nelze uložit.', 'error');
 					}
 				}
 
@@ -97,7 +102,7 @@
 					if (isset($existing)) {
 						$answer->set('existing_score', $existing->val('belbin_result_score'));
 					} else {
-						$answer->set('existing_score', 0);
+						$answer->set('existing_score', z::getInt('answer_' . $answer->val('belbin_answer_id'), 0));
 					}		
 				}
 				
@@ -106,6 +111,7 @@
 				$this->setData('test', $test);
 				$this->setData('question', $question);
 				$this->setData('answers', $answers);
+				$this->setData('form_token', $this->z->forms->createXSRFTokenHash('belbin_test_form'));
 				$this->setData('prev_question', $prev_question);
 				$this->insertJS(['score_per_question' => TestModel::$score_per_question]);
 				$this->includeJS('belbin.js', false);
